+++ .pre-commit-config.yaml
@@ -20,7 +20,7 @@ repos:
       rev: v0.1.6
       hooks:
           - id: ruff
-            args: [--fix, --exit-non-zero-on-fix]
+            args: [--fix, --exit-non-zero-on-fix, --unsafe-fixes]
           - id: ruff-format
     - repo: https://github.com/pre-commit/pre-commit-hooks
       rev: v4.5.0
