+++ pertpy/__init__.py
@@ -11,6 +11,7 @@ from numba import NumbaDeprecationWarning
 
 warnings.filterwarnings("ignore", category=NumbaDeprecationWarning)
 warnings.filterwarnings("ignore", category=MatplotlibDeprecationWarning)
+warnings.filterwarnings("ignore", category=UserWarning, module="scvi._settings")
 
 from . import data as dt
 from . import plot as pl
